+++ addons/portal/__manifest__.py
@@ -42,6 +42,7 @@ a dependency towards website editing and customization capabilities.""",
             'portal/static/src/js/portal_chatter.js',
             'portal/static/src/xml/portal_chatter.xml',
             'portal/static/src/js/portal_composer.js',
+            'portal/static/src/js/portal_security.js',
             'portal/static/src/js/portal_signature.js',
             'portal/static/src/xml/portal_signature.xml',
             'portal/static/src/js/portal_sidebar.js',
