+++ odoo/addons/base/__manifest__.py
@@ -69,6 +69,7 @@ The kernel of Odoo, needed for all installation.
         'views/res_country_views.xml',
         'views/res_currency_views.xml',
         'views/res_users_views.xml',
+        'views/res_users_identitycheck_views.xml',
         'views/ir_property_views.xml',
         'views/res_config_settings_views.xml',
         'views/report_paperformat_views.xml',
