+++ odoo/addons/base/views/res_users_identitycheck_views.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+
+        <record id="res_users_identitycheck_view_form" model="ir.ui.view">
+            <field name="model">res.users.identitycheck</field>
+            <field name="arch" type="xml">
+                <form string="Password Confirmation">
+                    <sheet class="bg-primary">
+                        <h3><strong>Please enter your password to confirm you own this account</strong></h3>
+                        <div>
+                            <field class="o_field_highlight col-10 col-md-6 px-0" name="password" autocomplete="current-password"
+                                required="True" password="True" placeholder="************"/>
+                        </div>
+                        <a href="/web/reset_password/" class="btn btn-link" role="button">Forgot password?</a>
+                    </sheet>
+                    <footer>
+                        <button string="Confirm Password" type="object" name="run_check" class="btn btn-primary" data-hotkey="q"/>
+                        <button string="Cancel" special="cancel" data-hotkey="z" class="btn btn-secondary"/>
+                    </footer>
+                </form>
+            </field>
+        </record>
+
+        <record id="res_users_identitycheck_view_form_revokedevices" model="ir.ui.view">
+            <field name="name">Revoke All Devices</field>
+            <field name="model">res.users.identitycheck</field>
+            <field name="priority">40</field>
+            <field name="arch" type="xml">
+                <form string="Log out from all devices">
+                    <div>
+                        You are about to log out from all devices that currently have access to your account.<br/><br/>
+                        <strong>Type in your password to confirm :</strong>
+                        <field class="oe_inline o_field_highlight" name="password" password="True" required="True"/>
+                    </div>
+                    <footer>
+                        <button string="Log out from all devices" name="revoke_all_devices" type="object" class="btn-primary" data-hotkey="q"/>
+                        <button string="Cancel" class="btn-secondary" special="cancel" data-hotkey="z" />
+                    </footer>
+                </form>
+            </field>
+        </record>
+
+    </data>
+</odoo>
